Changed lens_id from gdouble to gfloat.
git-svn-id: https://rawstudio.org/svn/rawstudio/branches/rawstudio-ng-color@2999 ff961bb0-e316-0410-bf66-ce93e2ff3b84

+++ librawstudio/rs-metadata.h
@@ -82,7 +82,7 @@ struct _RSMetadata {
 	GdkPixbuf *thumbnail;
 
 	/* Lens info */
-	gdouble lens_id;
+	gfloat lens_id;
 	gdouble lens_min_focal;
 	gdouble lens_max_focal;
 	gdouble lens_min_aperture;
